# Conflicts:
#	src/main/connect_front/src/member/Delete.jsx
#	src/main/connect_front/src/member/Edit.jsx
#	src/main/java/connectweb/connect_back/controller/member/MemberController.java
#	src/main/java/connectweb/connect_back/model/entity/board/BirthBoardEntity.java
#	src/main/java/connectweb/connect_back/model/entity/board/ReplyEntity.java
#	src/main/java/connectweb/connect_back/service/member/MemberService.java
#	src/main/resources/data.sql

+++ src/main/connect_front/src/member/Delete.jsx
@@ -1,28 +1,32 @@
 import axios from "axios";
-import { useRef, useState } from "react"
+import { useState } from "react"
 
-export default function Delete(){
+export default function Delete(props){
 
-    const [mPassword, setMpassword] = useState('');
+    const [mpw, setMpw] = useState('');
 
-    const onCheckPassword = async(e)=>{
-        setMpassword(e.target.value)
-        await axios.get("/conn/m/check.password", {params:{mpassword : mPassword}})
-        .then(response => {
-            console.log (response);
-            axios.delete("/conn/m/delete.do")
-            .then(response => {console.log (response)})
-            .catch(error=>{console.log(error)})
-        })
-        .catch(error=>{console.log(error)})
-
+    const onChangeMpw = (e)=>{
+        setMpw(e.target.value);
     }
 
+    const onDelete = async()=>{
+        await axios.get("/conn/m/check.password",{params:{mpw:mpw}})
+        .then(response=>{
+            if(response.data){
+                axios.delete("/conn/m/delete.do")
+                .then(response=>{
+                    alert('탈퇴 성공 이용해주셔서 감사합니다. GOOD BYE')
+                    window.location.href="/member/login"
+                })
+                .catch(error=>{console.log(error);})
+            }else{
+                alert('비밀번호가 일치하지 않습니다.')
+            }
+        })
+    }
 
     return(<>
-        <form ref={setDeleteFrom}>
-            <input type="text" name="mpassword" placeholder="비밀번호를 입력해주세요" onChange={onCheckPassword}/>
-            <button type="button" onClick={onDeleteMember}>탈퇴하기</button>
-        </form>
+        <input type="text" placeholder="비밀번호를 입력해주세요" value={mpw} onChange={onChangeMpw}/>
+        <button type="button" onClick={onDelete}>삭제하기</button>
     </>)
 }
